Updated tests; fixed transcription check
